Create a "game-ai" project for ai code
Updated the PlayerType so that it is no longer an enum but stores a list
of types.  Added a Service Provider so that the game-ai can register ai
players and game-core will see them and add them to the list of
PlayerType.

Added a stub for a new AI called flowfield that is in the game-ai
project.
