+++ game-ai/src/main/java/org/triplea/ai/flowfield/FlowFieldAi.java
@@ -1,0 +1,42 @@
+package org.triplea.ai.flowfield;
+
+import games.strategy.engine.data.GameData;
+import games.strategy.engine.data.GamePlayer;
+import games.strategy.engine.framework.startup.ui.PlayerTypes;
+import games.strategy.triplea.ai.AbstractAi;
+import games.strategy.triplea.delegate.remote.IAbstractPlaceDelegate;
+import games.strategy.triplea.delegate.remote.IMoveDelegate;
+import games.strategy.triplea.delegate.remote.IPurchaseDelegate;
+import games.strategy.triplea.delegate.remote.ITechDelegate;
+
+public class FlowFieldAi extends AbstractAi {
+  public FlowFieldAi(final String name, final PlayerTypes.AiType playerType) {
+    super(name, playerType);
+  }
+
+  @Override
+  protected void purchase(
+      final boolean purchaseForBid,
+      final int pusToSpend,
+      final IPurchaseDelegate purchaseDelegate,
+      final GameData data,
+      final GamePlayer player) {}
+
+  @Override
+  protected void tech(
+      final ITechDelegate techDelegate, final GameData data, final GamePlayer player) {}
+
+  @Override
+  protected void move(
+      final boolean nonCombat,
+      final IMoveDelegate moveDel,
+      final GameData data,
+      final GamePlayer player) {}
+
+  @Override
+  protected void place(
+      final boolean placeForBid,
+      final IAbstractPlaceDelegate placeDelegate,
+      final GameData data,
+      final GamePlayer player) {}
+}
