+++ game-ai/src/main/resources/META-INF/services/games.strategy.triplea.ai.AiProvider
@@ -1,0 +1,1 @@
+org.triplea.ai.flowfield.FlowFieldAiProvider
