+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/ServerSetupPanel.java
@@ -380,14 +380,15 @@ public class ServerSetupPanel extends SetupPanel implements IRemoteModelListener
           public void actionPerformed(final ActionEvent e) {
             if (enabledCheckBox.isSelected()) {
               model.enablePlayer(nameLabel.getText());
-              type.setSelectedItem(PlayerType.HUMAN_PLAYER);
+              type.setSelectedItem(PlayerTypes.HUMAN_PLAYER);
             } else {
               model.disablePlayer(nameLabel.getText());
-              type.setSelectedItem(PlayerType.WEAK_AI.name());
+              type.setSelectedItem(PlayerTypes.WEAK_AI);
             }
             setWidgetActivation();
           }
         };
+    private final PlayerTypes playerTypesProvider;
 
     PlayerRow(
         final String playerName,
@@ -402,7 +403,8 @@ public class ServerSetupPanel extends SetupPanel implements IRemoteModelListener
       enabledCheckBox.addActionListener(disablePlayerActionListener);
       // this gets updated later
       enabledCheckBox.setSelected(true);
-      final String[] playerTypes = PlayerType.playerTypes();
+      playerTypesProvider = new PlayerTypes();
+      final String[] playerTypes = playerTypesProvider.getAvailablePlayerLabels();
       type = new JComboBox<>(playerTypes);
       String previousSelection = reloadSelections.get(playerName);
       if (previousSelection.equalsIgnoreCase("Client")) {
@@ -411,11 +413,11 @@ public class ServerSetupPanel extends SetupPanel implements IRemoteModelListener
       if (!previousSelection.equals("no_one") && List.of(playerTypes).contains(previousSelection)) {
         type.setSelectedItem(previousSelection);
         model.setLocalPlayerType(
-            nameLabel.getText(), PlayerType.fromLabel((String) type.getSelectedItem()));
+            nameLabel.getText(), playerTypesProvider.fromLabel((String) type.getSelectedItem()));
       } else if (playerName.startsWith("Neutral") || playerName.startsWith("AI")) {
         // the 4th in the list should be Pro AI (Hard AI)
-        type.setSelectedItem(PlayerType.PRO_AI.getLabel());
-        model.setLocalPlayerType(nameLabel.getText(), PlayerType.PRO_AI);
+        type.setSelectedItem(PlayerTypes.PRO_AI.getLabel());
+        model.setLocalPlayerType(nameLabel.getText(), PlayerTypes.PRO_AI);
       }
       if (playerAlliances.contains(playerName)) {
         alliance = new JButton();
@@ -427,7 +429,8 @@ public class ServerSetupPanel extends SetupPanel implements IRemoteModelListener
       type.addActionListener(
           e ->
               model.setLocalPlayerType(
-                  nameLabel.getText(), PlayerType.fromLabel((String) type.getSelectedItem())));
+                  nameLabel.getText(),
+                  playerTypesProvider.fromLabel((String) type.getSelectedItem())));
     }
 
     public void takePlayerAction() {
