+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/posted/game/pbf/PbfSetupPanel.java
@@ -8,7 +8,7 @@ import games.strategy.engine.framework.startup.launcher.ILauncher;
 import games.strategy.engine.framework.startup.launcher.LocalLauncher;
 import games.strategy.engine.framework.startup.mc.HeadedLaunchAction;
 import games.strategy.engine.framework.startup.ui.PlayerSelectorRow;
-import games.strategy.engine.framework.startup.ui.PlayerType;
+import games.strategy.engine.framework.startup.ui.PlayerTypes;
 import games.strategy.engine.framework.startup.ui.SetupPanel;
 import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorModel;
 import games.strategy.engine.framework.startup.ui.posted.game.DiceServerEditor;
@@ -251,7 +251,7 @@ public class PbfSetupPanel extends SetupPanel implements Observer {
     }
 
     final PbemDiceRoller randomSource = new PbemDiceRoller(diceServerEditor.newDiceServer());
-    final Map<String, PlayerType> playerTypes = new HashMap<>();
+    final Map<String, PlayerTypes.Type> playerTypes = new HashMap<>();
     final Map<String, Boolean> playersEnabled = new HashMap<>();
     for (final PlayerSelectorRow player : this.playerTypes) {
       playerTypes.put(player.getPlayerName(), player.getPlayerType());
