+++ game-core/src/main/java/games/strategy/engine/player/Player.java
@@ -4,7 +4,7 @@ import games.strategy.engine.data.GamePlayer;
 import games.strategy.engine.data.Resource;
 import games.strategy.engine.data.Territory;
 import games.strategy.engine.data.Unit;
-import games.strategy.engine.framework.startup.ui.PlayerType;
+import games.strategy.engine.framework.startup.ui.PlayerTypes;
 import games.strategy.engine.message.IRemote;
 import games.strategy.engine.message.RemoteActionCode;
 import games.strategy.triplea.delegate.DiceRoll;
@@ -40,7 +40,7 @@ public interface Player extends IRemote {
   String getName();
 
   @RemoteActionCode(8)
-  PlayerType getPlayerType();
+  PlayerTypes.Type getPlayerType();
 
   /**
    * Start the given step. stepName appears as it does in the game xml file. The game step will
