+++ game-core/src/main/java/games/strategy/triplea/ai/pro/ProPurchaseAi.java
@@ -9,7 +9,7 @@ import games.strategy.engine.data.Territory;
 import games.strategy.engine.data.Unit;
 import games.strategy.triplea.Properties;
 import games.strategy.triplea.UnitUtils;
-import games.strategy.triplea.ai.AbstractAi;
+import games.strategy.triplea.ai.AbstractBuiltInAi;
 import games.strategy.triplea.ai.pro.data.ProBattleResult;
 import games.strategy.triplea.ai.pro.data.ProOtherMoveOptions;
 import games.strategy.triplea.ai.pro.data.ProPlaceTerritory;
@@ -2617,6 +2617,6 @@ class ProPurchaseAi {
         ProLogger.warn("Attempt was at: " + t + " with: " + unit);
       }
     }
-    AbstractAi.movePause();
+    AbstractBuiltInAi.movePause();
   }
 }
