+++ game-core/src/main/java/org/triplea/injection/Injections.java
@@ -1,6 +1,8 @@
 package org.triplea.injection;
 
 import com.google.common.base.Preconditions;
+import games.strategy.triplea.ai.AiProvider;
+import java.util.Collection;
 import lombok.Builder;
 import lombok.Getter;
 import org.triplea.util.Version;
@@ -68,6 +70,7 @@ public final class Injections {
   @Getter private static Injections instance;
 
   private final Version engineVersion;
+  private final Collection<AiProvider> aiProviders;
 
   public static synchronized void init(final Injections injections) {
     Preconditions.checkState(getInstance() == null);
