+++ game-core/src/test/java/games/strategy/engine/framework/startup/ui/PlayerTypesTest.java
@@ -1,6 +1,5 @@
 package games.strategy.engine.framework.startup.ui;
 
-import static java.util.Arrays.stream;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -11,27 +10,29 @@ import org.hamcrest.Matchers;
 import org.hamcrest.core.IsCollectionContaining;
 import org.junit.jupiter.api.Test;
 
-class PlayerTypeTest {
+class PlayerTypesTest {
 
   @Test
   void playerTypes() {
+    final PlayerTypes playerTypes = new PlayerTypes();
     assertThat(
         "Ensure we do not have an example invisible player type in the selection list",
-        List.of(PlayerType.playerTypes()),
-        Matchers.not(IsCollectionContaining.hasItem(PlayerType.CLIENT_PLAYER.getLabel())));
+        List.of(playerTypes.getAvailablePlayerLabels()),
+        Matchers.not(IsCollectionContaining.hasItem(PlayerTypes.CLIENT_PLAYER.getLabel())));
 
     assertThat(
         "Ensure we have a visible player type in the selection list",
-        List.of(PlayerType.playerTypes()),
-        IsCollectionContaining.hasItem(PlayerType.HUMAN_PLAYER.getLabel()));
+        List.of(playerTypes.getAvailablePlayerLabels()),
+        IsCollectionContaining.hasItem(PlayerTypes.HUMAN_PLAYER.getLabel()));
   }
 
   @Test
   void newPlayerWithName() {
     final String testName = "example";
 
-    stream(PlayerType.values())
-        .filter(playerType -> playerType != PlayerType.BATTLE_CALC_DUMMY)
+    final PlayerTypes playerTypesProvider = new PlayerTypes();
+    playerTypesProvider.getPlayerTypes().stream()
+        .filter(playerType -> !playerType.equals(PlayerTypes.BATTLE_CALC_DUMMY))
         .forEach(
             playerType -> {
               final Player result = playerType.newPlayerWithName(testName);
@@ -48,27 +49,33 @@ class PlayerTypeTest {
 
   @Test
   void fromLabel() {
-    assertThrows(IllegalStateException.class, () -> PlayerType.fromLabel("invalid_label_type"));
+    final PlayerTypes playerTypesProvider = new PlayerTypes();
+    assertThrows(
+        IllegalStateException.class, () -> playerTypesProvider.fromLabel("invalid_label_type"));
 
-    stream(PlayerType.values())
+    playerTypesProvider
+        .getPlayerTypes()
         .forEach(
             playerType ->
                 assertThat(
                     "Make sure that we can reconstruct each player type from its label",
-                    PlayerType.fromLabel(playerType.getLabel()),
+                    playerTypesProvider.fromLabel(playerType.getLabel()),
                     is(playerType)));
   }
 
   @Test
   void getLabel() {
+    final PlayerTypes playerTypes = new PlayerTypes();
     assertThat(
         "All player type labels should be unique, count of unique labels should match total",
-        stream(PlayerType.values()).map(PlayerType::getLabel).distinct().count(),
-        is((long) PlayerType.values().length));
+        playerTypes.getPlayerTypes().stream().map(PlayerTypes.Type::getLabel).distinct().count(),
+        is((long) playerTypes.getPlayerTypes().size()));
 
     assertThat(
         "No label should be empty ",
-        stream(PlayerType.values()).map(PlayerType::getLabel).anyMatch(String::isEmpty),
+        playerTypes.getPlayerTypes().stream()
+            .map(PlayerTypes.Type::getLabel)
+            .anyMatch(String::isEmpty),
         is(false));
   }
 }
