+++ game-headed/build.gradle
@@ -16,6 +16,7 @@ ext {
 dependencies {
     implementation "org.pushing-pixels:radiance-substance:$substanceVersion"
     implementation project(":game-core")
+    implementation project(":game-ai")
     implementation project(":java-extras")
     implementation project(":swing-lib")
     testImplementation "org.sonatype.goodies:goodies-prefs:$sonatypeGoodiesPrefsVersion"
