+++ settings.gradle
@@ -3,6 +3,7 @@ include 'database'
 include 'domain-data'
 include 'dropwizard-common'
 include 'dropwizard-common-test'
+include 'game-ai'
 include 'game-core'
 include 'game-headed'
 include 'game-headless'
